update tarefas, cliente, pagina inicial, report and login
