+++ frontend/paginaInicial/paginaInicialStyle.css
@@ -1,0 +1,108 @@
+body, html {
+    margin: 0;
+    padding: 0;
+    height: 100%;
+    font-family: Arial, sans-serif;
+}
+
+.container {
+    display: flex;
+    height: 100vh;
+    background-color: #0f6d70; /* cor de fundo */
+}
+
+.sidebar {
+    width: 20%;
+    background-color: #0f6d70; /* cor de fundo da barra lateral */
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.sidebar ul {
+    list-style-type: none;
+    padding: 0;
+    text-align: left;
+}
+
+.sidebar ul li {
+    margin: 20px 0;
+}
+
+.sidebar ul li a {
+    color: white;
+    text-decoration: none;
+    font-size: 20px;
+}
+
+.main-content {
+    width: 80%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    padding: 20px;
+    position: relative;
+}
+
+.logo {
+    margin-bottom: 20px;
+    margin-right: 26%;
+
+}
+
+.logo img {
+    max-width: 200px; /* ajuste conforme necessário */
+    display: block;
+    margin: 0 auto;
+}
+
+.news-section {
+    background-color: white;
+    padding: 20px;
+    border-radius: 10px;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    margin-left: 28%;
+
+}
+
+.news-section h2 {
+    margin: 0;
+    font-size: 24px;
+    margin-bottom: 20px;
+}
+
+.news-section img {
+    max-width: 100%;
+    height: auto;
+    border-radius: 10px;
+}
+
+/* Media Query para dispositivos móveis */
+@media screen and (max-width: 480px) {
+    .container {
+        flex-direction: column;
+    }
+
+    .sidebar {
+        width: 100%;
+        padding: 10px 0;
+    }
+
+    .main-content {
+        width: 100%;
+    }
+
+    .logo img {
+        max-width: 150px;
+    }
+
+    .news-section {
+        width: 90%;
+        top: 70%;
+    }
+}
