+++ frontend/paginaLogin/index.html
@@ -25,6 +25,10 @@
             <label>
                 <input type="password" placeholder="Senha" name="password" required>
             </label>
+            <label>
+                <input type="checkbox" id="showPassword"> Mostrar Senha
+            </label>
+
             <div class="remember-forgot">
                 <a href="#" id="forgotPasswordLink">Esqueceu a senha?</a>
             </div>
