+++ frontend/paginaMenuInicial/paginaMenuInicial.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Pagina Inicial</title>
-    <link rel="stylesheet" href="stylePaginaInicial.css">
+    <link rel="stylesheet" href="styleTarefas.css">
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 
 </head>
