+++ frontend/tarefas/erro.html
@@ -2,15 +2,15 @@
 <html>
 <head>
     <title>Erro nos Dados</title>
-    <meta http-equiv="refresh" content="3;url=/paginaInicial" />
+    <meta http-equiv="refresh" content="3;url=/tarefas" />
 </head>
 <body>
 <h1>Dados Inválidos</h1>
 <p>Você será redirecionado em 3 segundos.</p>
-<p>caso não tenha sido redirecionado em 3 segundos <a href="/paginaInicial">clique aqui</a></p>
+<p>caso não tenha sido redirecionado em 3 segundos <a href="/tarefas">clique aqui</a></p>
 <script>
     setTimeout(function() {
-        window.location.href = '/paginaInicial';
+        window.location.href = '/tarefas';
     }, 3000);
 </script>
 </body>
