+++ frontend/tarefas/paginaEditarTarefa/editarTarefa.html
@@ -9,7 +9,7 @@
 <button id="closeButton" onclick="fecharIframe()">X</button>
 <div id="alinhadorTarefa">
     <div id="tarefa">
-        <form id="editar-tarefa-form" action="/paginaInicial/editartarefa" method="post">
+        <form id="editar-tarefa-form" action="/tarefas/editartarefa" method="post">
             <label for="todo-input">Insira o novo nome da tarefa:</label>
             <input type="text" id="todo-input" name="titulo">
