+++ frontend/tarefas/styleTarefas.css
@@ -270,9 +270,12 @@ ul {
 }
 
 #taskPopup {
-    width: 300px;
-    box-shadow: 0 4px 8px rgba(0,0,0,0.1);
+    width: 35%;
+    max-height: 80vh; /* Define a altura máxima */
+    overflow-y: auto; /* Adiciona uma barra de rolagem vertical */
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
     color: #333; /* Cor escura para o texto */
+    padding: 20px;
 }
 
 #taskForm label {
@@ -280,24 +283,32 @@ ul {
     color: #333; /* Assegura que o texto do rótulo seja visível */
 }
 
-#taskForm input{
+#taskForm input,
+#taskForm textarea,
+#taskForm select {
     width: 100%;
     padding: 8px;
     box-sizing: border-box;
     background: #fff; /* Fundo branco para os campos de entrada */
     color: #333; /* Texto escuro para os campos de entrada */
     border: 1px solid #ccc; /* Borda sutil para os campos */
+    border-radius: 5px;
+    margin-bottom: 10px;
 }
 
 #taskForm textarea {
-    width: 100%; /* Garante que ocupe a largura total do formulário */
-    padding: 8px;
-    box-sizing: border-box;
-    background: #fff; /* Fundo branco para melhor visibilidade */
-    color: #333; /* Cor do texto */
-    border: 1px solid #ccc; /* Borda sutil */
-    resize: vertical; /* Permite redimensionar apenas verticalmente */
+    resize: vertical; /* Permitir redimensionamento vertical */
     min-height: 100px; /* Altura mínima inicial */
+}
+
+#taskForm button {
+    color: white; /* Texto branco para os botões */
+    padding: 10px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    margin-top: 10px;
+    width: 100%; /* Garante que os botões ocupem toda a largura do formulário */
 }
 
 .confirmarButtonTarefa {
@@ -306,6 +317,14 @@ ul {
 
 .confirmarButtonTarefa:hover {
     background-color: #45a049; /* Cor mais escura ao passar o mouse */
+}
+
+.cancelarButtonTarefa {
+    background-color: red;
+}
+
+.cancelarButtonTarefa:hover {
+    background-color: #7c0303;
 }
 
 #taskForm button {
@@ -533,6 +552,8 @@ select {
     font-size: 1.5rem;
     font-weight: 700;
     color: #333;
+    display: flex;
+    justify-content: center;
 }
 
 #taskDetailPopup label {
@@ -563,4 +584,115 @@ select {
 
 #taskDetailPopup button:hover {
     background-color: #45a049;
+}
+.desc-container {
+    display: flex;
+    align-items: center;
+    margin-top: 10px;
+}
+
+.desc-container textarea {
+    flex-grow: 1;
+    padding: 8px;
+    border-radius: 5px;
+    border: 1px solid #ccc;
+    margin-right: 10px;
+    resize: vertical; /* Permitir redimensionamento vertical */
+}
+
+.desc-container input[type="checkbox"] {
+    margin-right: 10px;
+}
+
+.desc-edit-icon,
+.desc-delete-icon {
+    cursor: pointer;
+    font-size: 1.2rem;
+    margin-left: 5px; /* Ajuste a margem conforme necessário */
+}
+
+.desc-delete-icon {
+    color: red;
+}
+
+#addDescriptionButton {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #4CAF50; /* Cor de fundo verde */
+    color: white;
+    border: none;
+    border-radius: 50%; /* Faz o botão circular */
+    width: 30px; /* Ajuste o tamanho conforme necessário */
+    height: 30px; /* Ajuste o tamanho conforme necessário */
+    cursor: pointer;
+    font-size: 1.5rem;
+    padding: 0; /* Remover qualquer padding adicional */
+    margin: 10px auto;
+}
+
+#addDescriptionButton:hover {
+    background-color: #45a049;
+}
+
+
+#toggleDescriptionTypeButton{
+    background-color: white;
+}
+
+.check-list-button{
+    margin-left: 98%; !important;
+}
+
+.custom-checkbox {
+    display: none;
+}
+
+.custom-checkbox + label {
+    display: inline-block;
+    position: relative;
+    padding-left: 30px;
+    cursor: pointer;
+    user-select: none;
+    font-size: 1rem;
+    line-height: 1.2;
+}
+
+.custom-checkbox + label:before,
+.custom-checkbox + label:after {
+    content: '';
+    position: absolute;
+    transition: all 0.3s ease;
+}
+
+.custom-checkbox + label:before {
+    left: 0;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 20px;
+    height: 20px;
+    border: 2px solid #ddd;
+    border-radius: 5px;
+    background: #fff;
+}
+
+.custom-checkbox:checked + label:before {
+    border-color: #09ad7e;
+    background-color: #09ad7e;
+}
+
+.custom-checkbox:checked + label:after {
+    content: '';
+    position: absolute;
+    top: 50%;
+    left: 5px;
+    width: 10px;
+    height: 5px;
+    border: solid #fff;
+    border-width: 0 2px 2px 0;
+    transform: translateY(-50%) rotate(134deg) rotateX(180deg);
+}
+
+.completed {
+    text-decoration: line-through;
 }
