+++ frontend/templateMenu/template.html
@@ -12,7 +12,7 @@
         <button class="hamburger-menu" onclick="toggleMenu()">☰</button>
         <ul id="menu" class="menu-inline">
             <li id="tarefas">
-                <a href="/paginaInicial"><button class="menu-item">Menu inicial</button></a>
+                <a href="/tarefas"><button class="menu-item">Tarefas</button></a>
             </li>
             <li>
                 <button class="menu-item" onclick="toggleDropdown('estudosDropdown')">Estudos</button>
